Updated database schema to add finer grained Config instances. Renamed compare_judgments to evaluate.
